requirements.txt
Flask
scikit-learn
numpy
gunicorn
pytesseract
Pillow
reportlab
